Completed
After some troubleshooting, we were able to run the code! We did not need the parenthesis next to calcTotal, because the function only being called at the event of onclick. Idk tbh.
The decimals in the 0's were removed because they're not actual values, it's just showing a value of nothing.

I have to make sure the script file is correct.